# docs/architecture/openapi-node.yaml
openapi: 3.0.3
info:
  title: DAgents Agent Node API
  version: "0.5"
  description: |
    Agent Node 本地 HTTP API。对外契约以 **Agent** 为中心（1 Agent = 1 主对话）。
    `/v1/sessions*` 已移除（未注册 → 404）；请使用 `/v1/agents/{agent_id}/...`。
servers:
  - url: http://127.0.0.1:18789
    description: 默认本机 Node
paths:
...
  /v1/triggers/{trigger_id}:
    delete:
      parameters:
        - name: trigger_id
          in: path
          required: true
          schema: {type: string}
        - name: revision
          in: query
          required: false
          schema: {type: integer, format: int64, minimum: 1}
          description: Expected revision for compare-and-swap deletion.
      x-contract-status: surface-only
      summary: Route registered in the Node HTTP server
      responses:
        "200":
          description: OK
        "400": {description: Invalid revision}
        "409": {description: Revision conflict or managed trigger}
    get:
      x-contract-status: surface-only
      summary: Route registered in the Node HTTP server
      responses:
        "200":
          description: OK
    patch:
      parameters:
        - name: trigger_id
          in: path
          required: true
          schema: {type: string}
      x-contract-status: surface-only
      summary: Route registered in the Node HTTP server
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TriggerUpdateRequest'}
      responses:
        "200":
          description: OK
        "400": {description: Invalid patch or revision}
        "409": {description: Revision conflict or managed trigger}
...
  /v1/triggers:
    get:
      x-contract-status: surface-only
      summary: Route registered in the Node HTTP server
      responses:
        "200":
          description: OK
    post:
      x-contract-status: surface-only
      summary: Route registered in the Node HTTP server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerCreateRequest'
      responses:
        "200":
          description: OK
...
  /v1/triggers/{trigger_id}/recover:
    post:
      summary: 丢弃指定的待投递记录并完成恢复
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerRecoveryRequest'
      responses:
        "200":
          description: 已恢复；触发器保持 disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerDefinition'
        "409":
          description: recovery_required、delivery_id 不匹配或无需恢复
      parameters:
        - name: trigger_id
          in: path
          required: true
          schema: {type: string}
...
components:
  schemas:
...
    TriggerRecoveryRequest:
      type: object
      required: [delivery_id]
      properties:
        delivery_id:
          type: string
          description: 必须与触发器当前 pending_delivery_id 完全匹配
    TriggerDefinition:
      type: object
      required: [trigger_id, name, condition, enabled]
      properties:
        trigger_id: {type: string}
        name: {type: string}
        condition: {type: object, additionalProperties: true}
        target_agent_id: {type: string}
        target_session_id: {type: string, nullable: true}
        session_target_mode: {type: string, enum: [fixed]}
        client_id: {type: string, nullable: true}
        task_template: {type: string}
        enabled: {type: boolean}
        pending_delivery_id: {type: string, nullable: true}
        pending_session_id: {type: string, nullable: true}
        recovery_required: {type: boolean}
        recovery_reason: {type: string}
        owner_agent_id: {type: string}
        controller: {type: string, enum: [user, goal, maintenance]}
        controller_id: {type: string}
        revision: {type: integer, format: int64}
        created_by: {type: string}
    TriggerCreateRequest:
      type: object
      required: [name, condition, task_template, target_agent_id]
      properties:
        name: {type: string}
        condition: {type: object, additionalProperties: true}
        target_agent_id: {type: string}
        target_session_id: {type: string, nullable: true}
        session_target_mode: {type: string, enum: [fixed]}
        task_template: {type: string}
        enabled:
          type: boolean
          default: true
    TriggerUpdateRequest:
      type: object
      properties:
        revision: {type: integer, format: int64, minimum: 1}
        name: {type: string}
        condition: {type: object, additionalProperties: true}
        target_agent_id: {type: string}
        target_session_id: {type: string, nullable: true}
        client_id: {type: string, nullable: true}
        task_template: {type: string}
        enabled: {type: boolean}
        session_target_mode: {type: string, enum: [fixed]}
...